이력서 등록 (이미지 제외)
DB에 내용 제대로 들어가는지 확인 완료. 아직 이미지 업로드는 처리하지 않음

+++ ONEFIT/src/main/java/org/zerock/b01/domain/trainer/Trainer.java
@@ -47,7 +47,7 @@ public class Trainer extends BaseEntity {
     private String wantLegion;
 
     @Column(nullable = false)
-    private int wantTime;
+    private Double wantTime;
 
     @Column(nullable = false)
     private int wantDay;
